+++ src/main/scala/com/dataintuitive/viash/functionality/dataobjects/DoubleObject.scala
@@ -21,7 +21,7 @@ case class DoubleObject(
   name: String,
   alternatives: List[String] = Nil,
   description: Option[String] = None,
-  example: Option[String] = None,
+  example: Option[Double] = None,
   default: Option[Double] = None,
   required: Boolean = false,
   tag: Option[String] = None,
@@ -29,4 +29,20 @@ case class DoubleObject(
   multiple: Boolean = false,
   multiple_sep: Char = ':',
   oType: String = "double"
-) extends DataObject[Double]
+) extends DataObject[Double] {
+  def copyDO(
+    oType: String, 
+    name: String, 
+    alternatives: List[String],
+    description: Option[String],
+    example: Option[Double],
+    default: Option[Double],
+    required: Boolean,
+    direction: Direction,
+    tag: Option[String],
+    multiple: Boolean,
+    multiple_sep: Char
+  ): DataObject[Double] = {
+    copy(name, alternatives, description, example, default, required, tag, direction, multiple, multiple_sep, oType)
+  }
+}