+++ src/main/scala/com/dataintuitive/viash/functionality/dataobjects/FileObject.scala
@@ -23,7 +23,7 @@ case class FileObject(
   name: String,
   alternatives: List[String] = Nil,
   description: Option[String] = None,
-  example: Option[String] = None,
+  example: Option[Path] = None,
   default: Option[Path] = None,
   must_exist: Boolean = false,
   required: Boolean = false,
@@ -32,4 +32,20 @@ case class FileObject(
   multiple: Boolean = false,
   multiple_sep: Char = ':',
   oType: String = "file"
-) extends DataObject[Path]
+) extends DataObject[Path] {
+  def copyDO(
+    oType: String, 
+    name: String, 
+    alternatives: List[String],
+    description: Option[String],
+    example: Option[Path],
+    default: Option[Path],
+    required: Boolean,
+    direction: Direction,
+    tag: Option[String],
+    multiple: Boolean,
+    multiple_sep: Char
+  ): DataObject[Path] = {
+    copy(name, alternatives, description, example, default, must_exist, required, tag, direction, multiple, multiple_sep, oType)
+  }
+}